requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
certifi==2022.6.15
chromedriver-autoinstaller==0.4.0
et-xmlfile==1.1.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
lxml==4.9.1
numpy==1.23.2
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-html-reporter==0.2.9
python-dateutil==2.8.2
python-dotenv==0.14.0
pytz==2022.2.1
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
xdist==0.0.2